Exercises 4.20-4.23 Blog List

+++ part4/bloglist/controllers/blogs.js
@@ -4,14 +4,6 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 const config = require('../utils/config')
 
-const getTokenFrom = (request) => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { blogs: 0 })
     response.json(blogs)
@@ -30,14 +22,15 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
-    const decodedToken = jwt.verify(getTokenFrom(request), config.SECRET)
+    const decodedToken = jwt.verify(request.token, config.SECRET)
     if (!decodedToken.id) {
         return response.status(401).json({
             error: 'token invalid'
         })
     }
 
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
+    console.log(user, "here is user")
 
     const blog = new Blog({
         title: body.title,
@@ -56,7 +49,26 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, config.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({
+            error: 'invalid token'
+        })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() !== decodedToken.id) {
+        return response.status(401).json({
+            error: 'invalid user'
+        })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
+
+    const user = request.user
+    console.log(user, "here is user")
+    user.blogs.pull(request.params.id);
+    await user.save();
 
     response.status(204).end()
 })
